Add support for Raspberry Pi Connect 1.3.0
Raspberry Pi Connect 1.3.0 adds a new rpi-connect-wayvnc-watcher.path
systemd unit that needs to be enabled so that rpi-connectd is informed
of the WayVNC service starting and stopping. It also makes any dependent
services "PartOf" the main rpi-connect.service so we no longer need to
explicitly stop or start them individually.

As rpi-connect 1.2.2 is included in the July 4th 2024 Raspberry Pi OS
image, we support both new and old versions, keeping the old logic for
stopping and starting rpi-connectd and its related services.