bootloader: Add an option to control when the network-install UI is displayed
The bootloader factory default configuration has been changed to display
the network-install UI for a few seconds after a cold boot. This is
enabled if NET_INSTALL_AT_POWER_ON is set to 1 in the bootloader
config (default is 0).

Add a menu option to control to disable (or re-enable) this setting.